Fixed windows file path issues and fixed clone and create new repo not working with spaces

+++ packages/rapid-cmi5/src/lib/design-tools/course-builder/GitViewer/utils/fileSystem.ts
@@ -1344,15 +1344,22 @@ export class GitFS {
       // Ensure the old file/folder exists
       const resp = await this.fs.promises.stat(fullOldPath);
 
+      // Needed for windows who does not care about caps
+      const isCaseOnlyRename =
+        fullOldPath.toLowerCase() === fullNewPath.toLowerCase() &&
+        fullOldPath !== fullNewPath;
+
       // Perform the rename (move)
       if (resp.isFile()) {
         const content = await this.fs.promises.readFile(fullOldPath);
         await this.fs.promises.writeFile(fullNewPath, content);
-        await this.fs.promises.rm(fullOldPath);
+        if (!isCaseOnlyRename) await this.fs.promises.rm(fullOldPath);
       } else {
         await this.copyRecursive(fullOldPath, fullNewPath);
-        await this.clearDirectory(fullOldPath);
-        await this.fs.promises.rmdir(fullOldPath);
+        if (!isCaseOnlyRename) {
+          await this.clearDirectory(fullOldPath);
+          await this.fs.promises.rmdir(fullOldPath);
+        }
       }
     } catch (error) {
       console.error('Error renaming file or folder:', error);
